msm: mdss: allocate new cursor buf based on FB_CUR_SETIMAGE
Cursor buffer is allocated and the img data is copied only
when there is a cursor size change. This might cause issues
when the cursor img changes retaining the same size. Update
the cursor buffer based on size change and also based on
the flag FB_CUR_SETIMAGE, which is set only when there is a
change in img data.

Change-Id: Iaf7aabd147b5fd1c2d72d26709c982601b3623da
Signed-off-by: Veera Sundaram Sankaran <veeras@codeaurora.org>
Signed-off-by: int0x19 <int0x19@gmail.com>